Add test_stepping.py
