+++ numba_dppy/tests/debugging/test_debug_dppy_numba.py
@@ -107,61 +107,3 @@ def test_local_variables(app):
 
     app.whatis("l1")
     app.child.expect(r"type = float64")
-
-
-# commands/next
-def test_next(app):
-    app.breakpoint("simple_dppy_func.py:30")
-    app.run("simple_dppy_func.py")
-
-    app.child.expect(r"Thread .* hit Breakpoint .* at simple_dppy_func.py:30")
-    app.child.expect(
-        r"30\s+c_in_kernel\[i\] = func_sum\(a_in_kernel\[i\], b_in_kernel\[i\]\)"
-    )
-
-    app.next()
-    app.next()
-
-    app.child.expect(r"Done\.\.\.")
-
-
-# commands/step_dppy_func
-def test_step(app):
-    app.breakpoint("simple_dppy_func.py:30")
-    app.run("simple_dppy_func.py")
-
-    app.child.expect(r"Thread .* hit Breakpoint .* at simple_dppy_func.py:30")
-    app.child.expect(
-        r"30\s+c_in_kernel\[i\] = func_sum\(a_in_kernel\[i\], b_in_kernel\[i\]\)"
-    )
-
-    app.step()
-    app.step()
-
-    app.child.expect(r"__main__::func_sum \(.*\) at simple_dppy_func.py:23")
-    app.child.expect(r"23\s+result = a_in_func \+ b_in_func")
-
-
-# commands/stepi
-def test_stepi(app):
-    app.breakpoint("simple_dppy_func.py:30")
-    app.run("simple_dppy_func.py")
-
-    app.child.expect(r"Thread .* hit Breakpoint .* at simple_dppy_func.py:30")
-    app.child.expect(
-        r"30\s+c_in_kernel\[i\] = func_sum\(a_in_kernel\[i\], b_in_kernel\[i\]\)"
-    )
-
-    app.stepi()
-
-    app.child.expect(
-        r"0x[0-f]+\s+30\s+c_in_kernel\[i\] = func_sum\(a_in_kernel\[i\], b_in_kernel\[i\]\)"
-    )
-
-    app.stepi()
-
-    app.child.expect(r"Switching to Thread")
-    app.child.expect(r"Thread .* hit Breakpoint .* at simple_dppy_func.py:30")
-    app.child.expect(
-        r"30\s+c_in_kernel\[i\] = func_sum\(a_in_kernel\[i\], b_in_kernel\[i\]\)"
-    )
